fix: nasty cache collision if two custom objects shared the same hash and type but where not ``equal``.
And as a matter of fact, this happens. For instance, all instance of
``object`` or any subclass will inherit a special ``hash`` method that
uses ``id``, but in some version of python (the recent ones), the ``id``
value is divided by ``16``. And hash collisions are to be expected
anyway, and of course should not cause cache collisions.